+++ src/hooks/index.ts
@@ -4,4 +4,5 @@ export type {ChartDetails} from './useLinearXAxisDetails';
 export {useLinearXScale} from './useLinearXScale';
 export {usePrevious} from './use-previous';
 export {useResizeObserver} from './useResizeObserver';
+export {useTheme} from './useTheme';
 export {usePolarisVizContext} from './usePolarisVizContext';