+++ CHANGELOG.md
@@ -5,6 +5,14 @@
 ### Fixed
 
 - Fixes `<BarChart />` when very large datasets are displayed
+### Added
+- `PolarisVizProvider` to support theming charts
+- `theme` to `BarChartProps`
+### Removed 
+- `barOptions`, `gridOptions`, `xAxisOptions.showTicks`, `xAxisOptions.labelColor` and `yAxisOptions.labelColor` from `BarChartProps`.
+
+### Changed
+- BarChart styles now are defined through themes in `PolarisVizProvider` instead of props. For more details check the [migration guide](https://docs.google.com/document/d/1VxfcgBbTNwjmYix1jGuDMgqDgIdehTgQbVZpER7djeU/edit?usp=sharing)
 
 ## [0.17.2] - 2021-06-23